refactor: app/adapter/src/main/scala/myapp/adapter/account/TransactionDto.scala -> app/adapter/src/main/scala/myapp/adapter/query/TransactionDto.scala
